FEATURE: Add flowQuery operations `referenceNodes` and `backReferenceNodes`
Other than the operations `reference` and `backReference` those return the nodes and not the reference edges.

This allows queries like the following example which returns all documents that have similar categories than the current one:
```
 ${q(documentNode).referenceNodes('categories').backReferenceNodes('categories').remove(documentNode).unique().get()}
```

!!! The new operations replace the `nodes` operation that which is removed. This is non breaking as `nodes` was never part of a release !!!
